+++ app/forms/transfer/task/rpa_policy_task_form.rb
@@ -1,0 +1,3 @@
+class Transfer::Task::RpaPolicyTaskForm < BaseTaskForm
+  attribute :confirm
+end